+++ src/structs/DefifaLaunchProjectData.sol
@@ -53,6 +53,6 @@ struct DefifaLaunchProjectData {
     IJBTerminal terminal;
     IJB721TiersHookStore store;
     uint256 minParticipation;
-    uint32 scorecardTimeout;
+    uint256 scorecardTimeout;
     uint256 timelockDuration;
 }
